Move Container class to raid package and tidy it up
Remove dead attribute nick and get_raw_array_size which is really
only appropriate for RAIDn classes and was renamed to get_net_array_size
in a previous commit.

Make Container.get_recommended_stride() raise an exception as the
notion of a recommended stride is meaningless for containers.
Do not catch the exception because if there's a situation where a Container
is being formatted, the only reason why get_recommended_stride() would be
called, something has gone wrong.

Add get_max_spares() and get_base_member_size() which both raise an exception.

Signed-off-by: mulhern <amulhern@redhat.com>