Make BTRFS devices use RAID objects instead of strings for levels
Update some comments on btrfs.create_volume() to match.

Also, refine BTRFSVolumeDevice._removeParent() method.
There is a change in behavior, if there is any RAIDlevel then the minimum
number of parents can never be less than 1. Previously, in the case of
single raid, it could be 0.

Make a separate list for btrfs's notion of a metadata RAID level.

Signed-off-by: mulhern <amulhern@redhat.com>