Add a flag to control whether a degraded md raid array is used (#1090009)
A degraded md raid array may still have enough disks to function.
The flag is set to True by default. Wrt. to the installer, it is generally
False, since installing to a degraded array is a bit reckless, but True
if the installer is in rescue mode, where it is desirable to locate as
many devices as possible.

Related: rhbz#1090009

Signed-off-by: mulhern <amulhern@redhat.com>