Add a property indicating whether a device is directly accessible.
Directly accessible means there are no layers of abstraction/indirection
between the device and a user. An example of a non-direct device is an LVM
PV that is part of a VG. An example of a direct device is a partition that
contains swap space.

It used to be the case that direct was equivalent to isleaf, but no more.
Actual leaf devices, all btrfs subvolumes (including the top-level), and
snapshot source/origin devices can all be accessed directly by users.