Automate generation of the .rst files which just set up the modules.
Since these .rst files were added a couple of files have moved or been
added and these .rst files should have been updated, but weren't. Now
they'll be updated automatically, and they won't be in the repository.

This isn't fully automatic, but the step of removing these particular *.rst
files is in the correct direction.

Signed-off-by: mulhern <amulhern@redhat.com>