Omit special check for md devices in addUdevDevice().
If the device appears to be an md device, just call addUdevMDDevice directly.

Use udev_device_get_md_uuid() to get the mduuid of the array, but catch
a key error. This has the same effect as the previous code, but is a bit
more proper, even if more wordy.

Signed-off-by: mulhern <amulhern@redhat.com>