Make DevicelibsTestCase devices configurable.
Allow choosing size of backing file. Do not use the path of the backing
file in all subclasses in previous strange way, since it is irrelevant
to all the subclasses, and only relevant to DevicelibsTestCase which has
to do the necessary cleanup. Use a new attribute, loopDevices, to store
the actual loopDevices, which are what is required by the subclasses.
Fix up the subclasses to use this attribute appropriately. Change directory
of backing file to /var/tmp instead of /tmp.

Signed-off-by: mulhern <amulhern@redhat.com>