Remove preferLeaves parameter from getDeviceByPath()
The only use of this parameter was removed in anaconda commit
8eb4f8f82ace112e7f216b6e70633a3549949187 (Wed Aug 8 09:58:51 2012).
It was added in anaconda commit 718622ed2b54979b09712c10bc94fd86f6ea81e8
(Mon Nov 21 09:50:25 2011 -0600).

This is a semantic change, the new version returns the first match it finds,
whereas the old version prefers a non-leaf match by default.
However, the list of devices is searched in order, and the list is generally
ordered with leaves at the end of the list, so a non-leaf device will generally
be found first.

Signed-off-by: mulhern <amulhern@redhat.com>