Canonicalize mdadm generated UUIDS (#1070095)
Related: rhbz#1070095

Examined alternatives are:
1) To make use of the uuid.UUID class to store uuids. The difficulty
with this idea is that some things that are labeled uuids are not proper
uuids at all.
2) To make the udev_info dict more object oriented, and have it sanitize
uuids when they are added. A lot of work for very limited benefit.
3) To canonicalize uuids everytime they are extracted from the info.
Brittle, since you can't stop code from accessing the entries directly,
and annoyingly inefficient.

This alternative is pretty simple, and to a certain extent it's a
workaround for a bug in mdadm, so it wins.

Signed-off-by: mulhern <amulhern@redhat.com>