Add a DiskFile class for testing partitioning code as a non-root user.
There seems to be some special treatment from parted (first free sector
is reported as 32) but the main purpose is to test partitioning code
without committing the changes to disk. It may be perfectly viable to
commit to these devices, but I haven't tried it yet. My intention is to
write unit tests for things like partition allocation -- which require
parted Device and Disk instances -- without requiring root access.