Replace our pyudev with the package python-pyudev.
There are two major differences in how we use the module. Each device
is represented by a pyudev.Device instance, which is based on
collections.abc.Mapping. That means we can no longer change the data
to do things like set a "multipath_member" format type or inject LVM
or MD metadata.

The other difference is that sysfs paths now include the "/sys", which
is great since it's always mounted in the same place.