Make has_redundancy() a method rather than a property and revise mdadd.
Originally, has_redundancy was set to False for Container so that it
could be a simple property. But this feels like a cheap trick, since
whether or not a Container level has intrinsic redundancy is not
a reasonable question. Now, raise an exception for Container.has_redundancy.

Change calling code appropriately. making sure that the exception is not
raised by that code and also that reasonably behavior occurs.

mdadd revisions:

1) LinearRAID mdadm call will fail if raid_devices is passed, so do not pass
raid_devices.
2) RAID0 mdadm call will fail if the the raid_devices number is not exactly
right, so use mddetail to find the current value and increment that by one.

Change calls accordingly.