Add two functions to enable manual addition of ECKD DASDs.
This just adds a couple of functions which aid in manually adding DASDs.
sanitize_dasd_dev_input takes a device number as a string and verifies
whether its format is acceptable as a match against DASD device numbers.

online_dasd takes a device number as a string and fill remove the device
from the ignore list and attempt to bring it online.