Remove poolMetaData (#1021505)
It is never anything but 0, and its only use is an addition or subtraction
and 0 is the identity for these operations.

It was added in commit 8ef070b7b6fe4f270cc0120e5425299abf90b8cd (2013/10/07).
Its last set was removed in commit 6897f50181662d91d1b6c84e749afe3ec48d2310
(2013/11/1).

It does not seem to have ever been used in anaconda, and it does not exist
in pykickstart.

Signed-off-by: mulhern <amulhern@redhat.com>