Add an epoch to blivet.
Because of an error made during branching, the first build of f21 blivet
incremented y-stream to 0.62 instead of z-stream to 0.61.1 and was
distributed to fedora users testing f21 alpha composes. This can't
easily be worked around any other way than by incrementing epoch,
otherwise 0.62 will always be seen as the "newer" version of blivet in
f21.

Also, because this change is going into f21, it must also go into master
branch.