Fix treatment of percent as lvm lv size spec.
Percent is an alternative to a literal size. It is a fixed size request
expressed as a percentage of the vg's free space after all requests with
an explicit size have been allocated (but not grown). That means that a
request defined using percent can still be growable. Such a request will
use the size calculated from the percent argument as the basis for growth
rate.

Related: rhbz#1116435