Disable LVM autobackup when doing image installs (#1066004)
If we're doing an image install, the host system doesn't really care
about the metadata for these volume groups - and if you do a lot of
image installs (say, on a build system) the archive/ directory is going
to grow without bound.

So: suppress the autobackup behavior during image installs to prevent
modifying things on the host.