Raise an exception when we find orphan partitions.
If we find a partition that belongs on a disk that we did not find a
disklabel on (or a hidden format or an iso9660 fs) we have serious
enough problems to justify raising an exception instead of quietly
ignoring it. Ignoring these partitions almost always ends up causing
fatal problems further along.

Related: rhbz#1144410