Update test setup so that it obeys RAID level requirements.
Previously, the MDRaidArray test did not check whether the number of
parents was fewer than the minimum number allowed by the level if there
were no parents at all. New code always checks, so all those tests that
managed to slip past by setting parents to [] or None have to be updated
with at least as many parents as are required by the level. The tests
that check the state of parents and devices must be updated along with
the setup.

Signed-off-by: mulhern <amulhern@redhat.com>