Abstract ContainerDevice member format check into a method (#1150147)
Resolves: fed#1150147

Override the member format check in MDRaidArrayDevice.

MDRaidArrayDevices are ContainerDevices that generally enforce
that the format of member devices is 'mdmember'. This is usually correct,
but devices which are MDRaidArrayDevices which themselves have type
mdcontainer may be members of MDRaidArrayDevices and they do not have
format of 'mdmember'. In this case, therefore, no restriction on
format type is enforced by MDRaidArrayDevice.

Signed-off-by: mulhern <amulhern@redhat.com>