Register DeviceFormat class (#1150147)
Related: fed#1150147

This shouldn't change any observable behavior in existing code.

The only behavioral change is that in getFormat(), the True branch of
if not fmt_class:
will be taken less often because get_device_format_class returns
DeviceFormat instead of None with an argument of None (assuming
DeviceFormat can be found).

This means that
if not fmt_class:
will evaluate to True only when a match can not be found, presumably
because its class could not be read properly.

Of course, if it's DeviceFormat that is not found, falling back to
DeviceFormat makes no sense, but things are pretty much over by then,
anyway.

Signed-off-by: mulhern <amulhern@redhat.com>