Take care when checking relationship of parent and child UUIDs (#1151649)
Related: rhbz#1151649

We are in the unfortunate position of not being able to assume that
the member's format has an attribute w/ the name specified by
self._formatUUIDAttr. This problem arises because MDRaidArrayDevices w/
type mdcontainer lack some of the properties of more typical
MDRaidArrayDevices.

Abstract all this checking into a single method, so that the problem doesn't
pop up again later when removing or adding (for real, not just in the
in memory model).

Preserve all semantics, except for not raising an AttributeError if the
field does not exist.

Distinguish between situation where the data is untrustworthy or
unavailable and where it is considered reliable enough that verifying
its correctness makes sense.

Don't assume that the problem can only be due to an oddity of mdcontainer,
it may crop up elsewhere as well, so be robust.

Signed-off-by: mulhern <amulhern@redhat.com>