Add xlate parameter to humanReadable().
Make default True, so no visible change.
We don't often look at logs generated in a foreign locale, but if we
did, and the locale were different enough from English, I would have
a bit of trouble interpreting the Size values since Size.__str__()
just turns around and invokes humanReadable(). We should probably,
for purposes of logs, explicitly invoke humanReadable(xlate=False).
This just sets up humanReadable() so that is possible.

Also, translate less eagerly.

The only prefix or abbreviation that needs to be translated is the
one that is displayed, so wait until display time to translate, if required.

Signed-off-by: mulhern <amulhern@redhat.com>