Try to deactivate lvm on corrupted gpt disks.
Since pyparted won't let us use the corrupt gpt disklabel we can't
treat the disk as partitioned, which prevents us from correctly detecting
the lvm on top of it. By deactivating the lvm when this situation arises,
we at least make it possible to clear the disk and do a fresh installation
on it.

Related: rhbz#1123450