Raise UnusableConfigurationError when unusable configuration is detected.
The two cases included initially are a corrupt gpt disklabel and
multiple lvm volume groups with the same name.

Blivet cannot proceed in the corrupt gpt case because pyparted doesn't
support instantiating a parted.Disk if the underlying gpt disklabel is
corrupt.

The lvm case has several issues, the most basic being that lvm is not
designed to handle multiple volume groups with the same name. A user
would have to rename one of the same-named vgs manually in order to
make it possible to use both vgs.

Resolves: rhbz#1123450