Don't use parted.Device to obtain size info.
Deleting a parted.Device causes parted to close its rw fd for the device,
which triggers a change uevent on that device, which could in turn trigger
any number of actions via udev rules. One example is when we reset a Blivet
instance: All devices are deleted, including their parted.Device instances,
which triggers a change uevent for every device. In response to these
events, mdadm's udev rules activate all arrays on those devices. Activating
devices -- even indirectly -- without cause, is not acceptable behavior for
a storage library.

This also adds a trailing comma to 1-tuples in variable_copy arguments.
In my testing, ('foo') is the string 'foo' -- not a tuple with lone element
'foo'.

Related: rhbz#1069597

Followup for removal of parted.Device.