Don't use parted.Device to obtain size info.
In order to see updated size info when a device is resize from outside of
blivet, it is necessary to delete the parted.Device and make another one.
This is happening from a change uevent handler. Deleting the parted.Device
causes parted to close its rw fd for the device, which triggers another
change uevent. We would have to ignore some of these change events to
avoid an infinite loop.

Something similar happens when resetting a Blivet instance. All devices
are deleted, including their parted.Device instances. This triggers change
uevents, which causes mdadm's udev rules to activate all arrays on those
devices, which is obviously not something we want to be triggering on a
reset. This one seems impossible to avoid as long as we're using the
parted.Device.

Related: rhbz#1069597