Move mediaPresent out of Device and into StorageDevice.
mediaPresent indicates whether a device node is backed by
readable/writeable media. It is generally only meaningful for root devices
(devices with no parents). There are only a few instances in which this
is relevant: cdrom/dvd, card readers, md containers, and some goofy raid
hardware that creates a zero-size block device node. If you need a general-
purpose property to tell you if a device is available for general use, what
you want is the device's status attribute.