Eliminate TmpFS.minSize()
TmpFS inherits both _minSize and _minInstanceSize from FS, where they are
both 0. TmpFS has no method to update _minInstanceSize, so _minInstanceSize
will always remain 0. Therefore, it would make sense for minSize() method
to just return the value of _minSize, indicating the the minimum size
for this filesystem is based just on its type, not on any information about
the particular instance. However, this is exactly what
DeviceFormat.minSize() which TmpFS would inherit if it didn't override,
does, so might as well just get rid of the overriding method, and let it
inherit.

This is a bit problematic, since 0 is used to mean unknown, but here means
0. It is possible that a method to obtain the minimum size of an existing
TmpFS instance should be written, but in the meantime, this change
makes the situation more obvious than it was previously.

Signed-off-by: mulhern <amulhern@redhat.com>