Run fsck before obtaining minimum filesystem size. (#1162215)
resize2fs -P now requires an e2fsck -f first. The lack of a real minimum
size left us with no useful minInstanceSize and allowed users to attempt
a shrink down to any size, which led to failures.

If we have a currentSize, set the default minInstanceSize to that so
that a failure to obtain a true minInstanceSize from the fs tools
precludes shrinking the filesystem.