Get rid of FS._getRandomUUID() method.
The method was introduced in anaconda commit
bd1123738d8a6ea671f472c663ed8e1c882f93a7 in 2011 Jan, but was not used in
that commit in anaconda or blivet and hasn't beeen used since.
It accompanied writeRandomUUID, which was removed in a previous commit.

If we really needed a random uuid method it should go in utils, but we do
not.

Signed-off-by: mulhern <amulhern@redhat.com>