Make selinux test less precise.
Previously they tested the exact value of the root_selinux_context.
However, when mounting and no default context available, the context
would not be set to the default value. So, the actual value of the
context is dependent on selinux's rules for the default context for that
root. It seems to have changed the rules for the default context for
directories in /tmp, now seeming to have no default context for those.

The real purpose of this test it to make sure that the lost+found directory
gets the correct context, when it exists. Those parts of the test are
doing just fine.

Signed-off-by: mulhern <amulhern@redhat.com>