Make logging a little less verbose and more useful in FS.mount()
Only log if there's a problem and make it clear what the problem is.

Previously logging was informational and could be deceptive,
in the first case claiming that the root SELinux context is set to None
(which never happens),
and in the second case asserting success even when failure had occurred.

Signed-off-by: mulhern <amulhern@redhat.com>