Simplify _getSize() and currentSize().
In _getSize():
Remove pointless boolean expression. Only returning self.targetSize if it
does not equal self._size is kind of pointless, because if they are equal
then it is just as good to return the value of one as the other.

Make currentSize() look more like _getSize() so they're easier to compare.

Signed-off-by: mulhern <amulhern@redhat.com>