Change allow_degraded_mdraid flag to allow_imperfect_devices (#1202877)
Related: rhbz#1090009
Related: rhbz#1202877

(cherry picked from commit 0eb434ba53d5c122205e56f7bad322780c3b65f8)

The only change is a name change. The plan is to use this flag for situations
where it is better to use the imperfect device then to deactivate or remove it
or otherwise get it out of sight.

Generally, when installing, imperfect devices are best avoided, because
you don't really want to install onto them. But when in installer rescue
mode and many other situations, it is better to expose an imperfect but
still usable device.

Signed-off-by: mulhern <amulhern@redhat.com>