Don't make overridden values actual properties.
When extending an abstract class which defines an abstract property it
is not necessary to make the overriding values properties, it's fine
just to make them regular class attributes. The abstract class checking
mechanism accepts these, and will have an error if they are not defined
at all.

The imaginary requirement to make them properties is something that seems
to follow naturally if you're thinking of a statically checked language.

It didn't cause any harm to make these fields properties, it was just
annoyingly verbose.

Signed-off-by: mulhern <amulhern@redhat.com>