Substitute simple value for single element array.
When you encounter an array, you kind of assume that there is some situation
under which it could contain other than exactly one value. If there is
no such situation, as is the case here, might as well not make it an
array. So, use format_designator instead of args.

Also, use format_designator in the error message. The error message is
about failure to look up a proper format, and it seems more reasonable
to include the format type with which the format was actually looked up,
which differs slightly from format_type.

Signed-off-by: mulhern <amulhern@redhat.com>