If allowing degraded array, attempt to start it (#1202877)
Resolves: rhbz#1090009
Related: rhbz#1202877

(cherry picked from commit ddb2e1756600d918df9df10384deb3364a9bbc3b)

If we can't get the name for the array, that's probably because the
array is in a state such that udevadm doesn't have the info. Try to
run the array and get the info again.

Signed-off-by: mulhern <amulhern@redhat.com>