Remove allowDirty parameter and code that depends on it.
If allowDirty is False then it is necessary to calculate the list of
dirty devices.

But, in the only place (in anaconda) where this method is invoked,
allowDirty is set to True.

Previously in anaconda, there were other places where mountExistingSystem()
was called, and sometimes allowDirty was False. But the last place this
occurred was in pyanaconda/upgrade.py, which was removed in commit
faea8ba2cf225974041c77dbe3977aebac91e880 (2012 Aug). I think that
means that now we can treat that parameter as if it is always True.

This patch requires a corresponding patch in anaconda (to remove use of
allowDirty parameter).

There is a tricky bit in the middle of the method where device.setup() is
called. I believe that this can be dropped from the main path of
mountExistingSystem() because setup() is called again in
FSSet.mountFilesystems().

Since this is the only place in which DirtyFSError is raised, remove that
definition as well.

Signed-off-by: mulhern <amulhern@redhat.com>