Remove needsFSCheck() and what only it depends on.
If needsFSCheck() is no longer invoked via mountExistingSystem() it is no
longer invoked anywhere. It is a bit of a dubious method because it depends
on values of self.dirty and self.errors which are set as a side-effect of
calling Ext2FS._getMinSize(). For the value of needsFSCheck() to
be meaningful, therefore, _getMinSize() must have been called and it must
have been called recently enough that the dirty and errors information is
not stale. It would be better to eliminate code which has an implicit
dependency on this timing behavior when it appears that the method and
supporting code has become unnecessary, anyway.

Signed-off-by: mulhern <amulhern@redhat.com>