Do not even pretend that ReiserFS is resizable.
To a resize request it will timidly respond:

resize_reiserfs 3.6.24

You are running BETA version of reiserfs shrinker.
This version is only for testing or VERY CAREFUL use.
Backup of you data is recommended.

Do you want to continue? [y/N]:

In man page, there is described a flag:

-f     Force, do not perform checks.

but no amount of -f'ing will persuade resize_reiserfs to just do it.

This is true not just in the currently distributed f20 version (3.6.21),
but also in the most recent version, 3.6.24.

ReiserFS is fairly moribund, I don't think this state of affairs is likely
to change.

Signed-off-by: mulhern <amulhern@redhat.com>