Change convertTo() and roundToNearest() so each takes a units specifier.
It's a lot nicer to pass them a named constant that they can use directly,
rather than a spec, which must be parsed as if it was entered by a user.

If there is ever a situation where the target of the conversion is user
specified, there is always _parseUnit() for converting a string to a unit.

Get rid of xlate parameter on convertTo(), it's pointless if the spec is
no longer being passed.

Change calling code appropriately.

This change requires a corresponding change in anaconda as convertTo() is
called there in two places.

The gains are clarity, simplicity, and efficiency.

Signed-off-by: mulhern <amulhern@redhat.com>