Use _resizefsUnit in resizeArgs() method implementations.
The reason it is used elsewhere, i.e., when calculating minimum size
is because of the constraint on the resizeArgs() method to use the
units defined by _resizefsUnit. This patch makes that explicit.

The appropriate format string for the size specification for the particular
resizing application and resizefsUnit combination is looked up dynamically
during execution of resizeArgs(). Ideally, it would be calculated when
the class is read, rather than at every invocation. However, due to the
fact that the format string is a function of the class's resizefsUnit
that is tricky. The only principled way to do it is to construct the class
dynamically, passing the resizefsUnit value as an argument. This is
definite overkill for this particular problem, so it seems better to
accept this slightly wasteful dynamic calculation.

There is no try/except KeyError block surrounding the FMT calculations.
Since the key is known statically and does not change between one invocation
and the next, this would be overkill.

Signed-off-by: mulhern <amulhern@redhat.com>